--- frontend/kobweb-compose/src/jsMain/kotlin/com/varabyte/kobweb/compose/ui/modifiers/UnicodeBidiModifiers.kt
@@ -1,9 +1,0 @@
-package com.varabyte.kobweb.compose.ui.modifiers
-
-import com.varabyte.kobweb.compose.css.*
-import com.varabyte.kobweb.compose.ui.Modifier
-import com.varabyte.kobweb.compose.ui.styleModifier
-
-fun Modifier.unicodeBidi(unicodeBidi: UnicodeBidi) = styleModifier {
-    unicodeBidi(unicodeBidi)
-}
